Move: SplashScreenのロジックをAppContainerに移動し、表示時間を調整
- スプラッシュスクリーンの初期化と管理をApp.tsxからAppContainerに移動。これにより、起動ロジックの一貫性と再利用性を向上。
- スプラッシュスクリーンの表示時間を調整してユーザー体験を改善。アプリケーションの起動がよりスムーズになるよう最適化。
