+++ src/components/AppContainer.tsx
@@ -1,6 +1,7 @@
 import { NavigationContainer } from '@react-navigation/native';
+import { hideAsync, preventAutoHideAsync } from 'expo-splash-screen';
 import { useColorMode, useTheme } from 'native-base';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { StatusBar } from 'react-native';
 import MainNavigator from './MainNavigator';
 
@@ -10,6 +11,20 @@ const AppContainer = (): JSX.Element => {
 
     const defaultDark = theme.colors.primary[900];
     const defaultLightThemeBgColor = 'white';
+
+    useEffect(() => {
+        const prepareResource = async () => {
+            try {
+                await preventAutoHideAsync();
+                await new Promise((resolve) => setTimeout(resolve, 1800));
+            } catch (e) {
+                console.error(e);
+            } finally {
+                await hideAsync();
+            }
+        };
+        prepareResource();
+    }, []);
 
     return (
         <>
